virtiofsd: validate path components
Several FUSE requests contain single path components.  A correct FUSE
client sends well-formed path components but there is currently no input
validation in case something went wrong or the client is malicious.

Refuse ".", "..", and paths containing '/' when we expect a path
component.

+++ contrib/virtiofsd/passthrough_ll.c
@@ -146,6 +146,21 @@ static void unref_inode(struct lo_data *lo, struct lo_inode *inode, uint64_t n);
 
 static struct lo_inode *lo_find(struct lo_data *lo, struct stat *st);
 
+static int is_dot_or_dotdot(const char *name)
+{
+	return name[0] == '.' && (name[1] == '\0' ||
+				  (name[1] == '.' && name[2] == '\0'));
+}
+
+/* Is `path` a single path component that is not "." or ".."? */
+static int is_safe_path_component(const char *path)
+{
+	if (strchr(path, '/')) {
+		return 0;
+	}
+
+	return !is_dot_or_dotdot(path);
+}
 
 static struct lo_data *lo_data(fuse_req_t req)
 {
@@ -657,6 +672,14 @@ static void lo_lookup(fuse_req_t req, fuse_ino_t parent, const char *name)
 		fuse_log(FUSE_LOG_DEBUG, "lo_lookup(parent=%" PRIu64 ", name=%s)\n",
 			parent, name);
 
+	/* Don't use is_safe_path_component(), allow "." and ".." for NFS export
+	 * support.
+	 */
+	if (strchr(name, '/')) {
+		fuse_reply_err(req, EINVAL);
+		return;
+	}
+
 	err = lo_do_lookup(req, parent, name, &e);
 	if (err)
 		fuse_reply_err(req, err);
@@ -718,6 +741,11 @@ static void lo_mknod_symlink(fuse_req_t req, fuse_ino_t parent,
 	struct lo_inode *dir;
 	struct fuse_entry_param e;
 	struct lo_cred old = {};
+
+	if (!is_safe_path_component(name)) {
+		fuse_reply_err(req, EINVAL);
+		return;
+	}
 
 	dir = lo_inode(req, parent);
 	if (!dir) {
@@ -815,6 +843,11 @@ static void lo_link(fuse_req_t req, fuse_ino_t ino, fuse_ino_t parent,
 	struct fuse_entry_param e;
 	int saverr;
 
+	if (!is_safe_path_component(name)) {
+		fuse_reply_err(req, EINVAL);
+		return;
+	}
+
 	inode = lo_inode(req, ino);
 	if (!inode) {
 		fuse_reply_err(req, EBADF);
@@ -854,6 +887,10 @@ out_err:
 static void lo_rmdir(fuse_req_t req, fuse_ino_t parent, const char *name)
 {
 	int res;
+	if (!is_safe_path_component(name)) {
+		fuse_reply_err(req, EINVAL);
+		return;
+	}
 
 	res = unlinkat(lo_fd(req, parent), name, AT_REMOVEDIR);
 
@@ -865,6 +902,12 @@ static void lo_rename(fuse_req_t req, fuse_ino_t parent, const char *name,
 		      unsigned int flags)
 {
 	int res;
+
+	if (!is_safe_path_component(name) ||
+		!is_safe_path_component(newname)) {
+		fuse_reply_err(req, EINVAL);
+		return;
+	}
 
 	if (flags) {
 		fuse_reply_err(req, EINVAL);
@@ -880,6 +923,11 @@ static void lo_rename(fuse_req_t req, fuse_ino_t parent, const char *name,
 static void lo_unlink(fuse_req_t req, fuse_ino_t parent, const char *name)
 {
 	int res;
+
+	if (!is_safe_path_component(name)) {
+		fuse_reply_err(req, EINVAL);
+		return;
+	}
 
 	res = unlinkat(lo_fd(req, parent), name, 0);
 
@@ -1031,12 +1079,6 @@ out_err:
 	fuse_reply_err(req, error);
 }
 
-static int is_dot_or_dotdot(const char *name)
-{
-	return name[0] == '.' && (name[1] == '\0' ||
-				  (name[1] == '.' && name[2] == '\0'));
-}
-
 static void lo_do_readdir(fuse_req_t req, fuse_ino_t ino, size_t size,
 			  off_t offset, struct fuse_file_info *fi, int plus)
 {
@@ -1178,6 +1220,11 @@ static void lo_create(fuse_req_t req, fuse_ino_t parent, const char *name,
 	if (lo_debug(req))
 		fuse_log(FUSE_LOG_DEBUG, "lo_create(parent=%" PRIu64 ", name=%s)\n",
 			parent, name);
+
+	if (!is_safe_path_component(name)) {
+		fuse_reply_err(req, EINVAL);
+		return;
+	}
 
 	err = lo_change_cred(req, &old);
 	if (err)
